+++ NEWS.md
@@ -4,6 +4,7 @@
 - `sd_question()` now has a `reactive` argument to enable reactive questions, and `sd_question_reactive()` was removed.
 - Made `sd_admin_enable()` internal, changed to `admin_enable()`.
 - Removed `sd_admin_ui()` and `sd_add_admin()` from the package.
+- Moved the surveydown.css and page_nav.js files into the R package and out of the Quarto extension.
 - Bug fixes: 
   - The admin page password was looking for the `SUPABASE_PASSWORD` environment variable, but it should be `SURVEYDOWN_PASSWORD`.
   - The data fetching was not working to download the data as a CSV file in the admin page, now uses `sd_get_data()` and works.